+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Major Changes
 * `RealmResults<T>` can be observed for granular changes via the new `SubscribeForNotifications` method.
+* `Realm` gained the `WriteAsync` method which allows a write transaction to be executed on a background thread.
 
 ### Breaking Changes
 * File format of Realm files is changed. Files will be automatically upgraded but opening a Realm file with older versions of Realm is not possible.